+++ lib/src/constants.dart
@@ -1,5 +1,6 @@
 class Constant {
   const Constant._();
+  static const int kDefaultMaxPages = 99999999;
   static const double kDefaultLineHeight = 1.0;
   static const int DEFAULT_FONT_SIZE = 12;
   static const String DEFAULT_FONT_FAMILY = 'Arial';